split up application methods
